extracted to a helper
