+++ Connectors/Rhino/Speckle.Connectors.RhinoShared/Operations/Receive/Settings/ToHostSettingsManager.cs
@@ -1,6 +1,5 @@
 using Speckle.Connectors.DUI.Models.Card;
 using Speckle.InterfaceGenerator;
-using Speckle.Sdk.Common;
 
 namespace Speckle.Connectors.Rhino.Operations.Receive.Settings;
 
@@ -10,6 +9,6 @@ public class ToHostSettingsManager : IToHostSettingsManager
   public bool GetConvertMeshesToBrepsSetting(ModelCard modelCard)
   {
     var value = modelCard.Settings?.FirstOrDefault(s => s.Id == ConvertMeshesToBrepsSetting.SETTING_ID)?.Value as bool?;
-    return value is not null && value.NotNull();
+    return value is true;
   }
 }
